requirements/production.txt
# Dependencies related to deploying woodshop
-r base.txt

# WSGI Handler
gevent==1.0.2
gunicorn==19.3.0

# Static and Media Storage
django-storages-redux==1.2.3

# Caching 
hiredis==0.2.0
django-redis-cache==0.13.1

# Django-secure
django-secure==1.0.1

# App Monitoring
newrelic==2.50.0.39
